Changed DatasetVersion.ProcessInstance foreign key to null on delete of the ProcessInstance record, rather than being deleted in the cascade.  This means that Datasts will not be removed during stream or task deletion.
The deleteTask and deleteStream methods will later have to be hooked
up to special code which handles the DatasetVersion deletion
correctly.  If a DatasetVersion is the only version for a dataset, the
master dataset record must also be removed.  If it's not the only
version, but it's the master, than a new master must be assigned.  Etc.

(GDCB-73)