Change select queries to include memidnum  in groups.jsp, modify.jsp, group.jsp, checkGroupsForUser.tag, mailList.jsp, user.jsp, userGroups.jsp, delete.jsp.
Make memidnum requried for "checkGroupsForUser".

In Modify.jsp:
Add email notification capability (not advertised yet. May switch to group-in-groups)
Include mailer2 and jstl functions in taglib

In Profile_UG insert memidnum as part of the key instead of user_id
 
In Group.jsp concatenate memidnum:username and pass as a single parameter to other jsps. They can split the argument to get the memidnum and userid.